+++ modules/apps/oauth2-provider/oauth2-provider-api/build.gradle
@@ -4,6 +4,5 @@ dependencies {
 	compileOnly group: "com.liferay", name: "com.liferay.portal.configuration.metatype", version: "2.0.0"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.6.0"
 	compileOnly group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
-	compileOnly group: "org.osgi", name: "org.osgi.service.metatype", version: "1.3.0"
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-scope-liferay-api")
 }
