+++ modules/apps/oauth2-provider/oauth2-provider-api/src/main/java/com/liferay/oauth2/provider/constants/GrantType.java
@@ -24,9 +24,8 @@ public enum GrantType {
 
 	AUTHORIZATION_CODE(true, true, false),
 	AUTHORIZATION_CODE_PKCE(true, false, true),
-	CLIENT_CREDENTIALS(false, true, false),
-	RESOURCE_OWNER_PASSWORD(false, true, true),
-	REFRESH_TOKEN(false, true, true);
+	CLIENT_CREDENTIALS(false, true, false), REFRESH_TOKEN(false, true, true),
+	RESOURCE_OWNER_PASSWORD(false, true, true);
 
 	public boolean isRequiresRedirectURI() {
 		return _requiresRedirectURI;
