+++ modules/apps/oauth2-provider/oauth2-provider-api/src/main/java/com/liferay/oauth2/provider/model/OAuth2AuthorizationModel.java
@@ -14,200 +14,11 @@
 
 package com.liferay.oauth2.provider.model;
 
-import com.liferay.expando.kernel.model.ExpandoBridge;
 import com.liferay.portal.kernel.model.BaseModel;
-import com.liferay.portal.kernel.model.CacheModel;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.util.HtmlUtil;
-
-import java.io.Serializable;
-
-import java.util.Date;
-import java.util.List;
-import java.util.Map;
 
 /**
  * @author Tomas Polesovsky
  */
-public class OAuth2AuthorizationModel
-	implements BaseModel<OAuth2Authorization> {
-
-	@Override
-	public Object clone() {
-		return null;
-	}
-
-	@Override
-	public int compareTo(OAuth2Authorization o) {
-		return 0;
-	}
-
-	@Override
-	public ExpandoBridge getExpandoBridge() {
-		return null;
-	}
-
-	@Override
-	public Map<String, Object> getModelAttributes() {
-		return null;
-	}
-
-	@Override
-	public Class<?> getModelClass() {
-		return null;
-	}
-
-	@Override
-	public String getModelClassName() {
-		return null;
-	}
-
-	@Override
-	public Serializable getPrimaryKeyObj() {
-		return null;
-	}
-
-	@Override
-	public boolean isCachedModel() {
-		return false;
-	}
-
-	@Override
-	public boolean isEntityCacheEnabled() {
-		return false;
-	}
-
-	@Override
-	public boolean isEscapedModel() {
-		return false;
-	}
-
-	@Override
-	public boolean isFinderCacheEnabled() {
-		return false;
-	}
-
-	@Override
-	public boolean isNew() {
-		return false;
-	}
-
-	@Override
-	public void resetOriginalValues() {
-	}
-
-	@Override
-	public void setCachedModel(boolean cachedModel) {
-	}
-
-	@Override
-	public void setExpandoBridgeAttributes(BaseModel<?> baseModel) {
-	}
-
-	@Override
-	public void setExpandoBridgeAttributes(ExpandoBridge expandoBridge) {
-	}
-
-	@Override
-	public void setExpandoBridgeAttributes(ServiceContext serviceContext) {
-	}
-
-	@Override
-	public void setModelAttributes(Map<String, Object> attributes) {
-	}
-
-	@Override
-	public void setNew(boolean n) {
-	}
-
-	@Override
-	public void setPrimaryKeyObj(Serializable primaryKeyObj) {
-	}
-
-	@Override
-	public CacheModel<OAuth2Authorization> toCacheModel() {
-		return null;
-	}
-
-	@Override
-	public OAuth2Authorization toEscapedModel() {
-		final OAuth2Authorization parent = (OAuth2Authorization)this;
-
-		OAuth2Authorization escapedModel = new OAuth2Authorization() {
-
-			@Override
-			public Date getAccessTokenExpirationDate() {
-				return parent.getAccessTokenExpirationDate();
-			}
-
-			@Override
-			public long getCompanyId() {
-				return parent.getCompanyId();
-			}
-
-			@Override
-			public Date getCreateDate() {
-				return parent.getCreateDate();
-			}
-
-			@Override
-			public long getoAuth2ApplicationId() {
-				return parent.getoAuth2ApplicationId();
-			}
-
-			@Override
-			public long getoAuth2RefreshTokenId() {
-				return parent.getoAuth2RefreshTokenId();
-			}
-
-			@Override
-			public long getoAuth2TokenId() {
-				return parent.getoAuth2TokenId();
-			}
-
-			@Override
-			public Date getRefreshTokenExpirationDate() {
-				return parent.getRefreshTokenExpirationDate();
-			}
-
-			@Override
-			public String getRemoteIPInfo() {
-				return HtmlUtil.escape(parent.getRemoteIPInfo());
-			}
-
-			@Override
-			public String getScopes() {
-				return HtmlUtil.escape(parent.getScopes());
-			}
-
-			@Override
-			public List<String> getScopesList() {
-				return parent.getScopesList();
-			}
-
-			@Override
-			public long getUserId() {
-				return parent.getUserId();
-			}
-
-			@Override
-			public String getUserName() {
-				return HtmlUtil.escape(parent.getUserName());
-			}
-
-		};
-
-		return escapedModel;
-	}
-
-	@Override
-	public OAuth2Authorization toUnescapedModel() {
-		return (OAuth2Authorization)this;
-	}
-
-	@Override
-	public String toXmlString() {
-		return toString();
-	}
-
+public interface OAuth2AuthorizationModel
+	extends BaseModel<OAuth2Authorization> {
 }
