+++ modules/apps/oauth2-provider/oauth2-provider-api/src/main/java/com/liferay/oauth2/provider/service/OAuth2ApplicationLocalServiceUtil.java
@@ -41,12 +41,6 @@ public class OAuth2ApplicationLocalServiceUtil {
 	 *
 	 * Never modify this class directly. Add custom service methods to {@link com.liferay.oauth2.provider.service.impl.OAuth2ApplicationLocalServiceImpl} and rerun ServiceBuilder to regenerate this class.
 	 */
-
-	/**
-	* NOTE FOR DEVELOPERS:
-	*
-	* Never reference this class directly. Always use {@link OAuth2ApplicationLocalServiceUtil} to access the o auth2 application local service.
-	*/
 	public static com.liferay.oauth2.provider.model.OAuth2Application addOAuth2Application(
 		long companyId, long userId, java.lang.String userName,
 		java.util.List<com.liferay.oauth2.provider.constants.GrantType> allowedGrantTypesList,
