+++ modules/apps/oauth2-provider/oauth2-provider-api/src/main/java/com/liferay/oauth2/provider/service/OAuth2RefreshTokenLocalService.java
@@ -82,6 +82,11 @@ public interface OAuth2RefreshTokenLocalService extends BaseLocalService,
 	public OAuth2RefreshToken createOAuth2RefreshToken(
 		long oAuth2RefreshTokenId);
 
+	/**
+	* NOTE FOR DEVELOPERS:
+	*
+	* Never reference this class directly. Always use {@link OAuth2RefreshTokenLocalServiceUtil} to access the o auth2 refresh token local service.
+	*/
 	public OAuth2RefreshToken createOAuth2RefreshToken(
 		java.lang.String tokenContent);
