+++ modules/apps/oauth2-provider/oauth2-provider-api/src/main/java/com/liferay/oauth2/provider/service/OAuth2RefreshTokenServiceUtil.java
@@ -41,6 +41,12 @@ public class OAuth2RefreshTokenServiceUtil {
 	 *
 	 * Never modify this class directly. Add custom service methods to {@link com.liferay.oauth2.provider.service.impl.OAuth2RefreshTokenServiceImpl} and rerun ServiceBuilder to regenerate this class.
 	 */
+
+	/**
+	* NOTE FOR DEVELOPERS:
+	*
+	* Never reference this class directly. Always use {@link OAuth2RefreshTokenServiceUtil} to access the o auth2 refresh token remote service.
+	*/
 	public static com.liferay.oauth2.provider.model.OAuth2RefreshToken deleteOAuth2RefreshToken(
 		long oAuth2RefreshTokenId)
 		throws com.liferay.portal.kernel.exception.PortalException {
