+++ modules/apps/oauth2-provider/oauth2-provider-service/src/main/java/com/liferay/oauth2/provider/model/impl/OAuth2ApplicationImpl.java
@@ -26,22 +26,11 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
- * The extended model implementation for the OAuth2Application service. Represents a row in the &quot;OAuth2Application&quot; database table, with each column mapped to a property of this class.
- *
- * <p>
- * Helper methods and all application logic should be put in this class. Whenever methods are added, rerun ServiceBuilder to copy their definitions into the {@link com.liferay.oauth2.provider.model.OAuth2Application} interface.
- * </p>
- *
  * @author Brian Wing Shun Chan
  */
 @ProviderType
 public class OAuth2ApplicationImpl extends OAuth2ApplicationBaseImpl {
 
-	/**
-	 * NOTE FOR DEVELOPERS:
-	 *
-	 * Never reference this class directly. All methods that expect a o auth2 application model instance should use the {@link com.liferay.oauth2.provider.model.OAuth2Application} interface instead.
-	 */
 	public OAuth2ApplicationImpl() {
 	}
